Inline media query style to let roadie ignore*
This issue happened before and I don't know what's the problem.

It turns out that, to inline the style Roadie can only make one set of
stylesheet worked, which is fair. After checking the html template,
https://github.com/leemunroe/responsive-html-email-template/blob/master/email-inlined.html,
it inlines the style and leaves the media query style in head. This commit
follows the behaviour.