Add fetchUser in index js*
The hard part here is how too connect the app between backend and frontend.
I was using sintra to render index.erb with hook code like

```js
// index.js
<script type="text/javascript" src="bundle.js"></script>
<script type="text/javascript">
  window.addEventListener("load", function() {
    window.Application.render(
      <%= view.to_json %>
    );
  });
</script>

// app.js
window.Application = {
  render: <App />
}
```

The problem here is that we lose the power of frontend dev environment,
by which I mean, I could not make webpack-dev-server to understand the
index.erb.

Right now, the approach changes from rendering data in erb to let
javascript request the data to render, which decouples the backend and
frontend code entirely. The index.html is a pure html file, understood
by both sinatra and webpack-dev-server. The flow is

1. Sinatra renders index.html which references the bundled js code
2. JS code fetches data evertime the index.html gets loaded

Hence, Sinatra works as the entrance of the app by rendering index.html
and also a API service to provide data to frontend.